Add new MCSkinHistory client
